refactor(066): move GameRoute into feature layer
